// src/analyses/interfaces/team.interface.ts
export interface Team {
  teamId: number;
  win: boolean;
  totalKills: number;
  totalAssists: number;
  totalDeaths: number;
  totalDamageDone: number;
  totalDamageTaken: number;
  totalHealed: number;
  avgKP: number;
  avgDeaths: number;
  avgDamageDone: number;
  avgDamageTaken: number;
  avgHealed: number;
}